+++ src/main/java/net/sf/jabref/logic/error/StreamEavesdropper.java
@@ -1,7 +1,15 @@
 package net.sf.jabref.logic.error;
 
-import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
+
+import javafx.application.Platform;
+
+import net.sf.jabref.logic.logging.LogMessages;
+
+import org.apache.logging.log4j.Level;
+import org.apache.logging.log4j.core.LogEvent;
+import org.apache.logging.log4j.core.impl.Log4jLogEvent;
+import org.apache.logging.log4j.message.SimpleMessage;
 
 /**
  * Allows to eavesdrop on an out and an err stream.
@@ -10,12 +18,14 @@ import java.io.PrintStream;
  */
 public class StreamEavesdropper {
 
-    private final ByteArrayOutputStream errByteStream = new ByteArrayOutputStream();
-    private final ByteArrayOutputStream outByteStream = new ByteArrayOutputStream();
-
     private final PrintStream systemOut;
     private final PrintStream systemErr;
 
+
+    public StreamEavesdropper(PrintStream systemOut, PrintStream systemErr) {
+        this.systemOut = systemOut;
+        this.systemErr = systemErr;
+    }
 
     public static StreamEavesdropper eavesdropOnSystem() {
         StreamEavesdropper streamEavesdropper = new StreamEavesdropper(System.out, System.err);
@@ -24,27 +34,50 @@ public class StreamEavesdropper {
         return streamEavesdropper;
     }
 
-    public StreamEavesdropper(PrintStream systemOut, PrintStream systemErr) {
-        this.systemOut = systemOut;
-        this.systemErr = systemErr;
-    }
-
+    /**
+     * Return a new {@code PrintStream} which also creates a new log event with {@link Level#WARN} for each message and forwards it to the {@link LogMessages} archive.
+     *
+     * @return a PrintStream
+     */
     public PrintStream getOutStream() {
-        PrintStream consoleOut = new PrintStream(outByteStream);
-        return new TeeStream(consoleOut, systemOut);
+        return new PrintStream(systemOut) {
+            @Override
+            public void write(byte[] buf, int off, int len) {
+                super.write(buf, off, len);
+                String message = new String(buf, off, len);
+                addToLog(message, Level.WARN);
+            }
+        };
     }
 
+    /**
+     * Return a new {@code PrintStream} which also creates a new log event with {@link Level#ERROR} for each message and forwards it to the {@link LogMessages} archive.
+     *
+     * @return a PrintStream
+     */
     public PrintStream getErrStream() {
-        PrintStream consoleErr = new PrintStream(errByteStream);
-        return new TeeStream(consoleErr, systemErr);
+        return new PrintStream(systemErr) {
+            @Override
+            public void write(byte[] buf, int off, int len) {
+                super.write(buf, off, len);
+                String message = new String(buf, off, len);
+                addToLog(message, Level.ERROR);
+            }
+        };
     }
 
-    public String getErrorMessages() {
-        return errByteStream.toString();
-    }
-
-    public String getOutput() {
-        return outByteStream.toString();
+    /**
+     * Creates a new log event with the given parameters and forwards it to the {@link LogMessages} archive.
+     *
+     * @param message message of log event
+     * @param level   level of log event
+     */
+    private void addToLog(String message, Level level) {
+        if (!message.equals(System.lineSeparator())) {
+            String messageFormat = message.replaceAll(System.lineSeparator(), "");
+            LogEvent messageWithLevel = Log4jLogEvent.newBuilder().setMessage(new SimpleMessage(messageFormat)).setLevel(level).build();
+            Platform.runLater(() -> LogMessages.getInstance().add(messageWithLevel));
+        }
     }
 
 }